backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
aiofiles==23.2.1
PyPDF2==3.0.1
python-docx==1.1.0
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.24.3
pandas==1.5.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
textstat==0.7.3
langdetect==1.0.9
websockets==12.0
httpx==0.25.0
python-dotenv==1.0.0
Authlib==1.2.0
itsdangerous==2.1.2
supabase
fpdf
unidecode
